requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
django-environ
psycopg2-binary
pillow
dotenv
django-debug-toolbar
